make event and job name more generic
