+++ backend/src/Designer/Events/DeploymentPipelineQueued.cs
@@ -3,9 +3,8 @@ using MediatR;
 
 namespace Altinn.Studio.Designer.Events;
 
-public class AppDecommissioningStarted : INotification
+public class DeploymentPipelineQueued : INotification
 {
     public required AltinnRepoEditingContext EditingContext { get; set; }
-    public required string Environment { get; set; }
     public required int BuildId { get; set; }
 }
